Add a tunnel manager so the data source can return and keep the tunnel running
The command that runs the tunnel blocks which means we need to run it in
a goroutine. Additionally, once all the provider resources are done
executing, Terraform appears to shut down the provider. We add a Depend
data source just so we can use `depends_on` in terraform to prevent a
premature shutdown.