Add a resource for the remote tunnel
This will allow creation of the tunnel in both the plan and apply
phases. When using a data source, only one phase will set up the tunnel
which isn't quite what we want.

In order to support the very first plan phase that is run, we add import
functionality (otherwise terraform would identify the resource as
needing to be created, but the create/read method wouldn't run and those
are needed to set up the tunnel).