Add refresh_id field for awsssmtunnels_remote_tunnel resource
We found that this resource doesn't always work with the plan phase
correctly. To fix this, we add a `refresh_id` field which will never
match what is read (which then will force an apply). This will make it
so the Read and Update methods run in plan and apply respectively,
setting up the tunnel.

This commit also references to the SSMRemoteTunnelResourceModel type.