Pass target to the provider configuration
Fix an issue where the plan phase uses the instance ID that is saved in
Terraform state. This instance may no longer exist and the Read method
doesn't utilize the provided attributes in HCL when setting up the
tunnel. This commit updates the schema for the Provider and Tunnel
resource so that the Target now is configured at the Provider level. To
utilize different targets, initialize separate providers.